requirements.txt
python-dotenv
syncsdk
vosk
sounddevice
numpy
piper-tts